add permissions for map

+++ app/Http/Controllers/MapsController.php
@@ -128,6 +128,15 @@ class MapsController extends Controller
 
     public function storeGu(Request $request): \Symfony\Component\HttpFoundation\Response
     {
+        if (!auth()->user()->hasPermissionTo('monitoring create gu', 'map-admin')) {
+            return response()->json(
+                [
+                    'status' => config('response.status.error'),
+                    'message' => trans('app.no_permissions_rights')
+                ]
+            );
+        }
+
         $gu_input = $request->input('gu');
         $gu = new Gu;
 
@@ -163,6 +172,15 @@ class MapsController extends Controller
 
     public function storeZu(Request $request): \Symfony\Component\HttpFoundation\Response
     {
+        if (!auth()->user()->hasPermissionTo('monitoring create zu', 'map-admin')) {
+            return response()->json(
+                [
+                    'status' => config('response.status.error'),
+                    'message' => trans('app.no_permissions_rights')
+                ]
+            );
+        }
+
         $zu_input = $request->input('zu');
         $zu = new Zu;
 
@@ -179,6 +197,15 @@ class MapsController extends Controller
 
     public function storeWell(Request $request, DruidService $druidService): \Symfony\Component\HttpFoundation\Response
     {
+        if (!auth()->user()->hasPermissionTo('monitoring create well', 'map-admin')) {
+            return response()->json(
+                [
+                    'status' => config('response.status.error'),
+                    'message' => trans('app.no_permissions_rights')
+                ]
+            );
+        }
+
         $well_input = $request->input('well');
         $well = new Well;
 
@@ -195,6 +222,15 @@ class MapsController extends Controller
 
     public function storePipe(Request $request): \Symfony\Component\HttpFoundation\Response
     {
+        if (!auth()->user()->hasPermissionTo('monitoring create pipe', 'map-admin')) {
+            return response()->json(
+                [
+                    'status' => config('response.status.error'),
+                    'message' => trans('app.no_permissions_rights')
+                ]
+            );
+        }
+
         $pipe_input = $request->input('pipe');
         $pipe = new OilPipe;
         $pipe->fill($pipe_input);
@@ -219,6 +255,15 @@ class MapsController extends Controller
 
     public function updateGu(Request $request, Gu $gu): \Symfony\Component\HttpFoundation\Response
     {
+        if (!auth()->user()->hasPermissionTo('monitoring update gu', 'map-admin')) {
+            return response()->json(
+                [
+                    'status' => config('response.status.error'),
+                    'message' => trans('app.no_permissions_rights')
+                ]
+            );
+        }
+
         $gu_input = $request->input('gu');
 
         $gu->fill($gu_input);
@@ -253,6 +298,15 @@ class MapsController extends Controller
 
     public function updateZu(Request $request, Zu $zu): \Symfony\Component\HttpFoundation\Response
     {
+        if (!auth()->user()->hasPermissionTo('monitoring update zu', 'map-admin')) {
+            return response()->json(
+                [
+                    'status' => config('response.status.error'),
+                    'message' => trans('app.no_permissions_rights')
+                ]
+            );
+        }
+
         $zu_input = $request->input('zu');
 
         $zu->fill($zu_input);
@@ -268,6 +322,15 @@ class MapsController extends Controller
 
     public function updateWell(Request $request, Well $well): \Symfony\Component\HttpFoundation\Response
     {
+        if (!auth()->user()->hasPermissionTo('monitoring update well', 'map-admin')) {
+            return response()->json(
+                [
+                    'status' => config('response.status.error'),
+                    'message' => trans('app.no_permissions_rights')
+                ]
+            );
+        }
+
         $well_input = $request->input('well');
 
         $well->fill($well_input);
@@ -283,6 +346,15 @@ class MapsController extends Controller
 
     public function updatePipe(Request $request, OilPipe $pipe): \Symfony\Component\HttpFoundation\Response
     {
+        if (!auth()->user()->hasPermissionTo('monitoring update pipe', 'map-admin')) {
+            return response()->json(
+                [
+                    'status' => config('response.status.error'),
+                    'message' => trans('app.no_permissions_rights')
+                ]
+            );
+        }
+
         $pipe_input = $request->input('pipe');
         $pipe->fill($pipe_input);
         $pipe->save();
@@ -305,6 +377,15 @@ class MapsController extends Controller
 
     public function deleteGu(Gu $gu): \Symfony\Component\HttpFoundation\Response
     {
+        if (!auth()->user()->hasPermissionTo('monitoring delete gu', 'map-admin')) {
+            return response()->json(
+                [
+                    'status' => config('response.status.error'),
+                    'message' => trans('app.no_permissions_rights')
+                ]
+            );
+        }
+
         $gu->delete();
 
         return response()->json(
@@ -316,6 +397,15 @@ class MapsController extends Controller
 
     public function deleteZu(Zu $zu): \Symfony\Component\HttpFoundation\Response
     {
+        if (!auth()->user()->hasPermissionTo('monitoring delete zu', 'map-admin')) {
+            return response()->json(
+                [
+                    'status' => config('response.status.error'),
+                    'message' => trans('app.no_permissions_rights')
+                ]
+            );
+        }
+
         $zu->delete();
 
         return response()->json(
@@ -327,6 +417,15 @@ class MapsController extends Controller
 
     public function deleteWell(Well $well): \Symfony\Component\HttpFoundation\Response
     {
+        if (!auth()->user()->hasPermissionTo('monitoring delete well', 'map-admin')) {
+            return response()->json(
+                [
+                    'status' => config('response.status.error'),
+                    'message' => trans('app.no_permissions_rights')
+                ]
+            );
+        }
+
         $well->delete();
 
         return response()->json(
@@ -338,6 +437,15 @@ class MapsController extends Controller
 
     public function deletePipe(OilPipe $pipe): \Symfony\Component\HttpFoundation\Response
     {
+        if (!auth()->user()->hasPermissionTo('monitoring delete pipe', 'map-admin')) {
+            return response()->json(
+                [
+                    'status' => config('response.status.error'),
+                    'message' => trans('app.no_permissions_rights')
+                ]
+            );
+        }
+
         PipeCoord::where('oil_pipe_id', $pipe->id)->delete();
         $pipe->delete();
